refactor(Visitor): rename the buildSchema file to Visitor
